refactor: enforce backend architecture boundaries
