--- backend/execution/repositories/controls.py
@@ -1,11 +1,0 @@
-from ...signals.repositories._store import (
-    get_effective_kill_switch,
-    list_kill_switch_events,
-    persist_kill_switch_event,
-)
-
-__all__ = [
-    "get_effective_kill_switch",
-    "list_kill_switch_events",
-    "persist_kill_switch_event",
-]
