+++ backend/market_data/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Pure market-data rules shared by ingestion and research consumers."""
