+++ backend/research/contracts/runs.py
@@ -31,11 +31,16 @@ from backend.shared.contracts.common import (
     StrategyType,
     ValidationMethod,
 )
-from backend.research.domain.artifact_summary import (
+from backend.research.contracts.artifacts import (
     ArtifactCompleteness,
     ReviewArtifactName,
 )
-from backend.research.domain.prospective_recipe import strict_recipe_issues
+from backend.research.policies.prospective import (
+    COHORT_2330,
+    COHORT_ALL_ACTIVE,
+    STRICT_MODE,
+    strict_recipe_issues,
+)
 from .runtime_metadata import (
     ConfigSources,
     EffectiveStrategyConfig,
@@ -117,8 +122,8 @@ class ValidationConfig(RequestModel):
 
 
 class ProspectiveEvidenceConfig(RequestModel):
-    mode: Literal["strict_v1"] = "strict_v1"
-    cohort_id: Literal["tw_2330_o2o_v1", "tw_all_active_o2o_v1"]
+    mode: Literal[STRICT_MODE] = STRICT_MODE
+    cohort_id: Literal[COHORT_2330, COHORT_ALL_ACTIVE]
     basis_date: date
     full_universe_symbols: conlist(str, min_length=1)  # type: ignore[valid-type]
