+++ backend/research/domain/opinion.py
@@ -3,13 +3,16 @@ from __future__ import annotations
 import math
 import re
 from collections.abc import Mapping
-from dataclasses import dataclass
-from datetime import date, datetime
 from typing import Any
 
 from backend.research.contracts.runs import (
     OpinionReviewCheckCategory,
     OpinionReviewCheckName,
+)
+from backend.research.domain.signal_snapshot import (
+    select_latest_signals as _latest_signals,
+    signal_date_key as _signal_date_key,
+    valid_latest_signals as _valid_latest_signals,
 )
 
 OPINION_ARTIFACT_VERSION = "phase2_opinion_artifact_v1"
@@ -30,14 +33,6 @@ BOUNDARY_RISK = (
     "Checked no-warning result: warning_count=0 and caveat_count=0; manual adoption "
     "review is still required."
 )
-
-
-@dataclass(frozen=True)
-class _LatestSignalSelection:
-    latest: list[Mapping[str, Any]]
-    invalid_row_count: int
-    unexpected_symbols: list[str]
-    snapshot_complete: bool
 
 
 def _as_mapping(value: Any) -> Mapping[str, Any]:
@@ -86,20 +81,6 @@ def _has_diagnostics_evidence(value: Any) -> bool:
             for field in ("rmse", "mae", "rank_ic", "linear_ic")
         )
     )
-
-
-def _signal_date_key(signal: Mapping[str, Any]) -> str | None:
-    value = signal.get("date")
-    if isinstance(value, datetime):
-        return value.date().isoformat()
-    if isinstance(value, date):
-        return value.isoformat()
-    if isinstance(value, str):
-        try:
-            return datetime.fromisoformat(value).date().isoformat()
-        except ValueError:
-            return None
-    return None
 
 
 def _refs(*items: tuple[str, str]) -> list[dict[str, str]]:
@@ -195,92 +176,6 @@ def _invalidation_context(payload: Mapping[str, Any]) -> tuple[str, list[dict[st
         [],
         "newer_run_or_market_data",
     )
-
-
-def _latest_signals(
-    payload: Mapping[str, Any],
-) -> _LatestSignalSelection:
-    declared = _as_list(payload.get("symbols")) or _as_list(
-        _as_mapping(payload.get("request_payload")).get("symbols")
-    )
-    allowed_symbols = {str(symbol) for symbol in declared if symbol}
-    forward_by_symbol: dict[
-        str, tuple[str, list[tuple[int, Mapping[str, Any]]]]
-    ] = {}
-    invalid_row_indices: set[int] = set()
-    unexpected_symbols: set[str] = set()
-    for row_index, item in enumerate(_as_list(payload.get("signals"))):
-        signal = _as_mapping(item)
-        if signal.get("signal_kind") != "forward_opinion":
-            continue
-        symbol = signal.get("symbol")
-        if not symbol:
-            invalid_row_indices.add(row_index)
-            continue
-        symbol_key = str(symbol)
-        if symbol_key not in allowed_symbols:
-            unexpected_symbols.add(symbol_key)
-            invalid_row_indices.add(row_index)
-            continue
-        signal_date = _signal_date_key(signal)
-        if signal_date is None:
-            invalid_row_indices.add(row_index)
-            continue
-        current = forward_by_symbol.get(symbol_key)
-        if current is None or signal_date > current[0]:
-            forward_by_symbol[symbol_key] = (
-                signal_date,
-                [(row_index, signal)],
-            )
-        elif signal_date == current[0]:
-            current[1].append((row_index, signal))
-
-    forward_dates = {selection[0] for selection in forward_by_symbol.values()}
-    common_as_of = max(forward_dates) if forward_dates else None
-    latest: list[Mapping[str, Any]] = []
-    for symbol in sorted(forward_by_symbol):
-        signal_date, candidates = forward_by_symbol[symbol]
-        selected_index, selected_signal = candidates[0]
-        latest.append(selected_signal)
-        invalid_row_indices.update(index for index, _ in candidates[1:])
-        if signal_date != common_as_of:
-            invalid_row_indices.add(selected_index)
-        if (
-            not _is_number(selected_signal.get("score"))
-            or not _is_number(selected_signal.get("position"))
-            or not _is_probability(selected_signal.get("up_probability"))
-            or selected_signal.get("predicted_direction") not in {"up", "down"}
-        ):
-            invalid_row_indices.add(selected_index)
-
-    # Keep structural invariants explicit even though invalid rows currently
-    # encode mixed dates and duplicate latest candidates.
-    snapshot_complete = (
-        bool(allowed_symbols)
-        and set(forward_by_symbol) == allowed_symbols
-        and len(forward_dates) == 1
-        and all(len(candidates) == 1 for _, candidates in forward_by_symbol.values())
-        and not invalid_row_indices
-    )
-    return _LatestSignalSelection(
-        latest=latest,
-        invalid_row_count=len(invalid_row_indices),
-        unexpected_symbols=sorted(unexpected_symbols),
-        snapshot_complete=snapshot_complete,
-    )
-
-
-def _valid_latest_signals(payload: Mapping[str, Any]) -> list[Mapping[str, Any]]:
-    latest = _latest_signals(payload).latest
-    return [
-        item
-        for item in latest
-        if _signal_date_key(item) is not None
-        and _is_number(item.get("score"))
-        and _is_number(item.get("position"))
-        and _is_probability(item.get("up_probability"))
-        and item.get("predicted_direction") in {"up", "down"}
-    ]
 
 
 def _action_rows(payload: Mapping[str, Any]) -> list[dict[str, Any]]:
