+++ backend/research/policies/__init__.py
@@ -1,0 +1,1 @@
+"""Pure research policies shared by contracts and application services."""
