--- backend/research/repositories/adaptive.py
@@ -1,19 +1,0 @@
-from ...signals.repositories._store import (
-    get_adaptive_profile,
-    list_adaptive_profiles,
-    list_adaptive_surface_exclusions,
-    list_adaptive_training_runs,
-    persist_adaptive_profile,
-    persist_adaptive_surface_exclusion,
-    persist_adaptive_training_run,
-)
-
-__all__ = [
-    "get_adaptive_profile",
-    "list_adaptive_profiles",
-    "list_adaptive_surface_exclusions",
-    "list_adaptive_training_runs",
-    "persist_adaptive_profile",
-    "persist_adaptive_surface_exclusion",
-    "persist_adaptive_training_run",
-]
