+++ tests/research/test_execution.py
@@ -7,9 +7,13 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
 import backend.research.services.execution as backtest_engine_service
-import backend.research.services.eligibility as eligibility_service
+import backend.market_data.services.research_inputs as eligibility_service
+from backend.market_data.repositories import official_audits as official_audits_repository
 from backend.database import Base, RawIngestAudit
-from backend.platform.errors import InsufficientDataError
+from backend.platform.errors import (
+    InsufficientDataError,
+    UnsupportedConfigurationError,
+)
 from backend.research.contracts.runs import (
     ResearchRunCreateRequest,
     ValidationConfig,
@@ -24,7 +28,11 @@ def _eligibility_session(monkeypatch):
         autoflush=False, autocommit=False, bind=engine
     )
     Base.metadata.create_all(bind=engine, tables=[RawIngestAudit.__table__])
-    monkeypatch.setattr(eligibility_service, "SessionLocal", testing_session_local)
+    monkeypatch.setattr(
+        official_audits_repository,
+        "SessionLocal",
+        testing_session_local,
+    )
     return testing_session_local
 
 
@@ -101,6 +109,40 @@ def test_research_eligibility_excludes_non_official_row_on_official_no_data(
     ).empty
 
 
+def test_research_eligibility_uses_inclusive_tw_start_date_floor(
+    monkeypatch,
+):
+    session_local = _eligibility_session(monkeypatch)
+    trading_date = date(2026, 7, 10)
+    fetch_timestamp_floor = datetime.combine(
+        trading_date,
+        datetime.min.time(),
+        tzinfo=eligibility_service.TW_TIMEZONE,
+    ).astimezone(timezone.utc)
+    with session_local() as session:
+        session.add_all(
+            [
+                _official_audit(
+                    source_name=source,
+                    trading_date=trading_date,
+                    payload_body='{"stat": "沒有符合條件"}',
+                    fetch_timestamp=fetch_timestamp,
+                )
+                for source in eligibility_service._BATCH_SOURCES
+                for fetch_timestamp in (
+                    fetch_timestamp_floor - timedelta(microseconds=1),
+                    fetch_timestamp_floor,
+                )
+            ]
+        )
+        session.commit()
+
+    assert eligibility_service.load_official_no_data_dates(
+        start_date=trading_date,
+        end_date=trading_date,
+    ) == {trading_date}
+
+
 def test_research_eligibility_excludes_audits_before_tw_start_date_floor(
     monkeypatch,
 ):
@@ -109,7 +151,7 @@ def test_research_eligibility_excludes_audits_before_tw_start_date_floor(
     fetch_timestamp_floor = datetime.combine(
         trading_date,
         datetime.min.time(),
-        tzinfo=eligibility_service.market_data_ingestion.TW_TIMEZONE,
+        tzinfo=eligibility_service.TW_TIMEZONE,
     ).astimezone(timezone.utc)
     with session_local() as session:
         session.add_all(
@@ -306,7 +348,11 @@ def test_research_eligibility_keeps_rows_when_audit_query_fails(monkeypatch, cap
         def execute(self, statement):
             raise RuntimeError("audit query unavailable")
 
-    monkeypatch.setattr(eligibility_service, "SessionLocal", _FailingSession)
+    monkeypatch.setattr(
+        official_audits_repository,
+        "SessionLocal",
+        _FailingSession,
+    )
 
     assert not eligibility_service.load_official_no_data_dates(
         start_date=date(2026, 7, 10), end_date=date(2026, 7, 10)
@@ -338,7 +384,7 @@ def test_research_eligibility_keeps_rows_when_audit_evaluation_fails(
         raise RuntimeError("audit evaluation unavailable")
 
     monkeypatch.setattr(
-        eligibility_service.market_data_ingestion,
+        eligibility_service.official_daily,
         "payload_declares_no_data",
         _fail_evaluation,
     )
@@ -395,6 +441,24 @@ def test_request_rejects_unshifted_daily_features():
 
     with pytest.raises(ValidationError):
         ResearchRunCreateRequest.model_validate(payload)
+
+
+def test_strict_prospective_features_reject_missing_model_columns():
+    unshifted = pd.DataFrame(
+        {"MA_5": [1.0]},
+        index=pd.to_datetime(["2024-01-04"]),
+    )
+
+    with pytest.raises(
+        UnsupportedConfigurationError,
+        match="peer_feature_value_p8",
+    ):
+        backtest_engine_service._build_prospective_prediction_features(
+            unshifted,
+            ["MA_5", "peer_feature_value_p8"],
+            symbol="2330",
+            strict=True,
+        )
 
 
 def test_direction_model_config_defaults_and_probability_bounds():
