+++ tests/shared/test_repository_fail_fast.py
@@ -82,8 +82,8 @@ def test_persist_repositories_fail_fast(monkeypatch, module, func_name, payload)
 @pytest.mark.parametrize(
     ("module", "func_name", "args"),
     [
-        (research_run_repository, "get_research_run_record", ("run_123",)),
-        (research_run_repository, "list_research_run_records", (20,)),
+        (research_run_repository, "get_research_run_snapshot", ("run_123",)),
+        (research_run_repository, "list_research_run_snapshots", (20,)),
         (replay_repository, "list_replay_records", (20,)),
         (recovery_repository, "list_recovery_records", (20,)),
         (lifecycle_repository, "list_lifecycle_records", (20,)),
